requirements.txt
fastapi
transformers
torch
torchvision
Pillow
uvicorn
python-multipart